Script to Add users
This is a script to add users to linux. You can also set passwords and the ssh keys of the created users will be copied to the location you set on the pc. Set the permission as 770 i.e enter command "chmod 770 /pathtofile/useraddsc.sh" and run the script by typing "bash /pathtofile/useradd.sc